Move the mock implementations in test directory
