+++ app/src/androidTest/java/com/github/fribourgsdp/radio/BuggyLobbyActivityTest.kt
@@ -11,7 +11,6 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.fribourgsdp.radio.mockimplementations.BuggyDatabase
 import com.github.fribourgsdp.radio.mockimplementations.BuggyLobbyActivity
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
