+++ app/src/androidTest/java/com/github/fribourgsdp/radio/LocalDatabaseTest.kt
@@ -2,7 +2,7 @@ package com.github.fribourgsdp.radio
 
 import android.content.ContentValues
 import android.util.Log
-import com.github.fribourgsdp.radio.mockimplementations.LocalDatabase
+import com.github.fribourgsdp.radio.LocalDatabase
 import com.google.android.gms.tasks.Tasks
 import org.junit.Test
 import java.util.concurrent.TimeUnit
